requirements.txt
aiohttp>=3.9.0
hidapi>=0.14.0
python-rtmidi>=1.5.0
